OSIS-461 Moved learning unit methods to business
