+++ base/business/learning_unit.py
@@ -25,8 +25,12 @@
 ##############################################################################
 import datetime
 from collections import OrderedDict
+
+from django.utils.translation import ugettext_lazy as _
+
 from base import models as mdl
 from base.business.learning_unit_year_with_context import volume_learning_component_year
+from base.forms.learning_units import LearningUnitYearForm
 from base.models import entity_container_year
 from base.models.entity_component_year import EntityComponentYear
 from base.models.entity_container_year import EntityContainerYear
@@ -38,11 +42,14 @@ from base.models.learning_container_year import LearningContainerYear
 from base.models.learning_unit import LearningUnit
 from base.models.learning_unit_component import LearningUnitComponent
 from base.models.learning_unit_year import LearningUnitYear
+from base.views.learning_unit import SIMPLE_SEARCH, SERVICE_COURSES_SEARCH
 from cms import models as mdl_cms
 from cms.enums import entity_name
 
 
 # List of key that a user can modify
+from osis_common.document import xls_build
+
 VALID_VOLUMES_KEYS = [
     'VOLUME_TOTAL',
     'VOLUME_Q1',
@@ -301,3 +308,60 @@ def create_learning_unit_year(academic_year, data, learning_container_year, lear
                                            quadrimester=data['quadrimester'])
 
 
+def _prepare_xls_content(found_learning_units):
+    if not found_learning_units:
+        return []
+    return [_extract_xls_data_from_learning_unit(lu) for lu in found_learning_units]
+
+
+def _extract_xls_data_from_learning_unit(learning_unit):
+    return [learning_unit.academic_year.name, learning_unit.acronym, learning_unit.title,
+            xls_build.translate(learning_unit.learning_container_year.container_type),
+            xls_build.translate(learning_unit.subtype),
+            _get_entity_acronym(learning_unit.entities.get('REQUIREMENT_ENTITY')),
+            _get_entity_acronym(learning_unit.entities.get('ALLOCATION_ENTITY')),
+            learning_unit.credits, xls_build.translate(learning_unit.status)]
+
+
+def _prepare_xls_parameters_list(request, workingsheets_data):
+    return {xls_build.LIST_DESCRIPTION_KEY: "Liste d'activités",
+            xls_build.FILENAME_KEY: 'Learning_units',
+            xls_build.USER_KEY:  _get_name_or_username(request.user),
+            xls_build.WORKSHEETS_DATA:
+                [{xls_build.CONTENT_KEY: workingsheets_data,
+                  xls_build.HEADER_TITLES_KEY: [str(_('academic_year_small')),
+                                                str(_('code')),
+                                                str(_('title')),
+                                                str(_('type')),
+                                                str(_('subtype')),
+                                                str(_('requirement_entity_small')),
+                                                str(_('allocation_entity_small')),
+                                                str(_('credits')),
+                                                str(_('active_title'))],
+                  xls_build.WORKSHEET_TITLE_KEY: 'Learning_units',
+                  }
+                 ]}
+
+
+def _get_name_or_username(a_user):
+    person = mdl.person.find_by_user(a_user)
+    if person:
+        return "{}, {}".format(person.last_name, person.first_name)
+    return a_user.username
+
+
+def _get_entity_acronym(an_entity):
+    return an_entity.acronym if an_entity else None
+
+
+def _get_search_form(request):
+    academic_year_id = request.GET.get('academic_year_id')
+    return LearningUnitYearForm(request.GET) if academic_year_id else LearningUnitYearForm()
+
+
+def _get_learning_units(form, search_type):
+    if search_type == SIMPLE_SEARCH:
+        return  form.get_activity_learning_units()
+    elif search_type == SERVICE_COURSES_SEARCH:
+        return form.get_service_course_learning_units()
+    return None
