+++ base/tests/views/test_learning_unit.py
@@ -34,7 +34,7 @@ from django.core.urlresolvers import reverse
 from django.http import HttpResponse, HttpResponseForbidden
 from django.test import TestCase, RequestFactory
 
-
+import base.business.learning_unit
 from base.forms import learning_units
 from base.forms.learning_units import CreateLearningUnitYearForm
 from base.models import learning_unit_component
@@ -783,17 +783,17 @@ class LearningUnitViewTestCase(TestCase):
     def test_get_username_with_no_person(self):
         a_username = 'dupontm'
         a_user = UserFactory(username=a_username)
-        self.assertEqual(learning_unit._get_name_or_username(a_user), a_username)
+        self.assertEqual(base.business.learning_unit._get_name_or_username(a_user), a_username)
 
     def test_get_username_with_no_person(self):
         a_user = UserFactory(username='dupontm')
         last_name='dupont'
         first_name='marcel'
         self.person = PersonFactory(user=a_user, last_name=last_name, first_name=first_name)
-        self.assertEqual(learning_unit._get_name_or_username(a_user), '{}, {}'.format(last_name, first_name))
+        self.assertEqual(base.business.learning_unit._get_name_or_username(a_user), '{}, {}'.format(last_name, first_name))
 
     def test_prepare_xls_content_no_data(self):
-        self.assertEqual(learning_unit._prepare_xls_content(None),[])
+        self.assertEqual(base.business.learning_unit._prepare_xls_content(None), [])
 
 class LearningUnitCreate(TestCase):
     def setUp(self):
