+++ base/views/learning_unit.py
@@ -42,14 +42,15 @@ from base.business.learning_unit import create_learning_unit, create_learning_un
     get_common_context_learning_unit_year, get_cms_label_data, \
     extract_volumes_from_data, get_same_container_year_components, get_components_identification, show_subtype, \
     get_organization_from_learning_unit_year, get_campus_from_learning_unit_year, \
-    get_all_attributions, get_last_academic_years
+    get_all_attributions, get_last_academic_years, _prepare_xls_content, _prepare_xls_parameters_list, _get_search_form, \
+    _get_learning_units
 from base.forms.common import TooManyResultsException
 from base.models import proposal_learning_unit, entity_version
 from base.models.campus import Campus
 from base.models.enums import learning_container_year_types, learning_unit_year_subtypes
 from base.models.enums.learning_unit_year_subtypes import FULL
 from base.models.learning_container import LearningContainer
-from base.forms.learning_units import LearningUnitYearForm, CreateLearningUnitYearForm, EMPTY_FIELD
+from base.forms.learning_units import CreateLearningUnitYearForm, EMPTY_FIELD
 from base.forms.learning_unit_specifications import LearningUnitSpecificationsForm, LearningUnitSpecificationsEditForm
 from base.forms.learning_unit_pedagogy import LearningUnitPedagogyForm, LearningUnitPedagogyEditForm
 from base.forms.learning_unit_component import LearningUnitComponentEditForm
@@ -460,60 +461,3 @@ def create_xls(request, found_learning_units):
     return xls_build.generate_xls(_prepare_xls_parameters_list(request, workingsheets_data))
 
 
-def _prepare_xls_content(found_learning_units):
-    if not found_learning_units:
-        return []
-    return [_extract_xls_data_from_learning_unit(lu) for lu in found_learning_units]
-
-
-def _extract_xls_data_from_learning_unit(learning_unit):
-    return [learning_unit.academic_year.name, learning_unit.acronym, learning_unit.title,
-            xls_build.translate(learning_unit.learning_container_year.container_type),
-            xls_build.translate(learning_unit.subtype),
-            _get_entity_acronym(learning_unit.entities.get('REQUIREMENT_ENTITY')),
-            _get_entity_acronym(learning_unit.entities.get('ALLOCATION_ENTITY')),
-            learning_unit.credits, xls_build.translate(learning_unit.status)]
-
-
-def _prepare_xls_parameters_list(request, workingsheets_data):
-    return {xls_build.LIST_DESCRIPTION_KEY: "Liste d'activités",
-            xls_build.FILENAME_KEY: 'Learning_units',
-            xls_build.USER_KEY:  _get_name_or_username(request.user),
-            xls_build.WORKSHEETS_DATA:
-                [{xls_build.CONTENT_KEY: workingsheets_data,
-                  xls_build.HEADER_TITLES_KEY: [str(_('academic_year_small')),
-                                                str(_('code')),
-                                                str(_('title')),
-                                                str(_('type')),
-                                                str(_('subtype')),
-                                                str(_('requirement_entity_small')),
-                                                str(_('allocation_entity_small')),
-                                                str(_('credits')),
-                                                str(_('active_title'))],
-                  xls_build.WORKSHEET_TITLE_KEY: 'Learning_units',
-                  }
-                 ]}
-
-
-def _get_name_or_username(a_user):
-    person = mdl.person.find_by_user(a_user)
-    if person:
-        return "{}, {}".format(person.last_name, person.first_name)
-    return a_user.username
-
-
-def _get_entity_acronym(an_entity):
-    return an_entity.acronym if an_entity else None
-
-
-def _get_search_form(request):
-    academic_year_id = request.GET.get('academic_year_id')
-    return LearningUnitYearForm(request.GET) if academic_year_id else LearningUnitYearForm()
-
-
-def _get_learning_units(form, search_type):
-    if search_type == SIMPLE_SEARCH:
-        return  form.get_activity_learning_units()
-    elif search_type == SERVICE_COURSES_SEARCH:
-        return form.get_service_course_learning_units()
-    return None
